Move AggregateOperationSupport to types.ts
